Refactor Google OAuth Logic: Move utility functions and OAuth client instantiation to backend
Related to #1332
